wasm: present GPL3 during configure
Present GPL3 as the open source license during configure
for Qt for WebAssembly. Commercial licensing is not
changed.

This inserts the “GPL3 or LGPL3” choice at the root of
the available licenses decision tree (the configure
script makes this choice, not the user). If LGPL3 is
chosen then we maybe add GPL2, as before.

On ‘g’ key press there is then no ambiguity and we
display either GPL2 or GPL3.

Change-Id: I8d2e3fc2df456e307d80a37e745afbcc902ff22b
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>