Don't set the wasDeleted flag in ~QWidget
Setting it too early can cause issues, as we are still doing
lots of work here as e.g. calling hide() on children.

But the flag is required when we delete the declarative data,
so set and reset it when destroying that data.

Amends c579f49e2a80a55a4004ff8e5b2ee76bda146387

Task-number: QTBUG-68637
Change-Id: I7ed35828c26912aa6d703ba6025e46b7911353fa
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>